refactor: Update profile screen layout and replace SafeButtonWidget with SaveButtonWidget
